Endre navn på harAlleredeSisteSats og flytt ut i util fil
