+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/autovedtak/satsendring/AutovedtakSatsendringService.kt
@@ -3,8 +3,6 @@ package no.nav.familie.ba.sak.kjerne.autovedtak.satsendring
 import io.micrometer.core.instrument.Metrics
 import no.nav.familie.ba.sak.common.Feil
 import no.nav.familie.ba.sak.common.UtbetalingsikkerhetFeil
-import no.nav.familie.ba.sak.common.isSameOrAfter
-import no.nav.familie.ba.sak.common.isSameOrBefore
 import no.nav.familie.ba.sak.config.TaskRepositoryWrapper
 import no.nav.familie.ba.sak.kjerne.autovedtak.AutovedtakService
 import no.nav.familie.ba.sak.kjerne.autovedtak.satsendring.domene.SatskjøringRepository
@@ -15,12 +13,8 @@ import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingStatus
 import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingType
 import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingÅrsak
 import no.nav.familie.ba.sak.kjerne.beregning.BeregningService
-import no.nav.familie.ba.sak.kjerne.beregning.SatsService
 import no.nav.familie.ba.sak.kjerne.beregning.TilkjentYtelseValidering
-import no.nav.familie.ba.sak.kjerne.beregning.domene.AndelTilkjentYtelseMedEndreteUtbetalinger
 import no.nav.familie.ba.sak.kjerne.beregning.domene.AndelerTilkjentYtelseOgEndreteUtbetalingerService
-import no.nav.familie.ba.sak.kjerne.beregning.domene.SatsType
-import no.nav.familie.ba.sak.kjerne.beregning.domene.YtelseType
 import no.nav.familie.ba.sak.kjerne.fagsak.FagsakStatus
 import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.PersongrunnlagService
 import no.nav.familie.ba.sak.kjerne.steg.StegType
@@ -171,7 +165,7 @@ class AutovedtakSatsendringService(
                 sisteIverksettBehandlingsId
             )
 
-        return harAlleredeSisteSats(andeler, satstidspunkt)
+        return andeler.erOppdatertMedSatserTilOgMed(satstidspunkt)
     }
 
     private fun harUtbetalingerSomOverstiger100Prosent(sisteIverksatteBehandling: Behandling): Boolean {
@@ -203,42 +197,6 @@ class AutovedtakSatsendringService(
     companion object {
         val logger = LoggerFactory.getLogger(AutovedtakSatsendringService::class.java)
         val secureLogger = LoggerFactory.getLogger("secureLogger")
-
-        fun harAlleredeSisteSats(
-            aty: List<AndelTilkjentYtelseMedEndreteUtbetalinger>,
-            satstidspunkt: YearMonth
-        ): Boolean {
-            val atyPåSatstidspunkt = aty.filter {
-                it.stønadFom.isSameOrBefore(satstidspunkt) && it.stønadTom.isSameOrAfter(
-                    satstidspunkt
-                )
-            }
-
-            val atySmåbarnstillegg = atyPåSatstidspunkt.filter { it.type == YtelseType.SMÅBARNSTILLEGG }
-            if (atySmåbarnstillegg.isNotEmpty()) {
-                val harGammelSats =
-                    atySmåbarnstillegg.any { it.sats != SatsService.finnSisteSatsFor(SatsType.SMA).beløp }
-                if (harGammelSats) return false
-            }
-
-            val atyUtvidet = atyPåSatstidspunkt.filter { it.type == YtelseType.UTVIDET_BARNETRYGD }
-            if (atyUtvidet.isNotEmpty()) {
-                val harGammelSats =
-                    atyUtvidet.any { it.sats != SatsService.finnSisteSatsFor(SatsType.UTVIDET_BARNETRYGD).beløp }
-                if (harGammelSats) return false
-            }
-
-            val atyOrdinær = atyPåSatstidspunkt.filter { it.type == YtelseType.ORDINÆR_BARNETRYGD }
-            if (atyOrdinær.isNotEmpty()) {
-                val satser = atyOrdinær.map { it.sats }
-                val gyldigeOrdinæreSatser = listOf(
-                    SatsService.finnSisteSatsFor(SatsType.ORBA).beløp,
-                    SatsService.finnSisteSatsFor(SatsType.TILLEGG_ORBA).beløp
-                )
-                satser.forEach { if (!gyldigeOrdinæreSatser.contains(it)) return false }
-            }
-            return true
-        }
     }
 }
