+++ src/test/enhetstester/kotlin/no/nav/familie/ba/sak/kjerne/autovedtak/satsendring/StartSatsendringTest.kt
@@ -12,7 +12,6 @@ import no.nav.familie.ba.sak.common.toYearMonth
 import no.nav.familie.ba.sak.config.FeatureToggleConfig
 import no.nav.familie.ba.sak.config.FeatureToggleService
 import no.nav.familie.ba.sak.config.TaskRepositoryWrapper
-import no.nav.familie.ba.sak.kjerne.autovedtak.satsendring.AutovedtakSatsendringService.Companion.harAlleredeSisteSats
 import no.nav.familie.ba.sak.kjerne.autovedtak.satsendring.domene.Satskjøring
 import no.nav.familie.ba.sak.kjerne.autovedtak.satsendring.domene.SatskjøringRepository
 import no.nav.familie.ba.sak.kjerne.behandling.domene.Behandling
@@ -236,17 +235,17 @@ internal class StartSatsendringTest {
         val atyMedBareSmåbarnstillegg =
             lagAndelTilkjentYtelseMedEndreteUtbetalinger(SatsType.SMA, behandling, SMÅBARNSTILLEGG)
 
-        assertThat(harAlleredeSisteSats(atyMedBareSmåbarnstillegg, SATSTIDSPUNKT)).isEqualTo(true)
+        assertThat(atyMedBareSmåbarnstillegg.erOppdatertMedSatserTilOgMed(SATSTIDSPUNKT)).isEqualTo(true)
 
         val atyMedBareUtvidet =
             lagAndelTilkjentYtelseMedEndreteUtbetalinger(SatsType.UTVIDET_BARNETRYGD, behandling, UTVIDET_BARNETRYGD)
 
-        assertThat(harAlleredeSisteSats(atyMedBareUtvidet, SATSTIDSPUNKT)).isEqualTo(true)
+        assertThat(atyMedBareUtvidet.erOppdatertMedSatserTilOgMed(SATSTIDSPUNKT)).isEqualTo(true)
 
         val atyMedBareOrba =
             lagAndelTilkjentYtelseMedEndreteUtbetalinger(SatsType.ORBA, behandling, ORDINÆR_BARNETRYGD)
 
-        assertThat(harAlleredeSisteSats(atyMedBareOrba, SATSTIDSPUNKT)).isEqualTo(true)
+        assertThat(atyMedBareOrba.erOppdatertMedSatserTilOgMed(SATSTIDSPUNKT)).isEqualTo(true)
 
         val atyMedBareTilleggOrba =
             lagAndelTilkjentYtelseMedEndreteUtbetalinger(
@@ -255,11 +254,10 @@ internal class StartSatsendringTest {
                 ORDINÆR_BARNETRYGD
             )
 
-        assertThat(harAlleredeSisteSats(atyMedBareTilleggOrba, SATSTIDSPUNKT)).isEqualTo(true)
+        assertThat(atyMedBareTilleggOrba.erOppdatertMedSatserTilOgMed(SATSTIDSPUNKT)).isEqualTo(true)
 
         assertThat(
-            harAlleredeSisteSats(
-                atyMedBareTilleggOrba + atyMedBareOrba + atyMedBareUtvidet + atyMedBareSmåbarnstillegg,
+            (atyMedBareTilleggOrba + atyMedBareOrba + atyMedBareUtvidet + atyMedBareSmåbarnstillegg).erOppdatertMedSatserTilOgMed(
                 SATSTIDSPUNKT
             )
         ).isEqualTo(true)
@@ -271,7 +269,7 @@ internal class StartSatsendringTest {
         val atyMedUgyldigSatsSmåbarnstillegg =
             lagAndelTilkjentYtelseMedEndreteUtbetalinger(SatsType.SMA, behandling, SMÅBARNSTILLEGG, UGYLDIG_SATS)
 
-        assertThat(harAlleredeSisteSats(atyMedUgyldigSatsSmåbarnstillegg, SATSTIDSPUNKT)).isEqualTo(false)
+        assertThat(atyMedUgyldigSatsSmåbarnstillegg.erOppdatertMedSatserTilOgMed(SATSTIDSPUNKT)).isEqualTo(false)
 
         val atyMedUglydligSatsUtvidet =
             lagAndelTilkjentYtelseMedEndreteUtbetalinger(
@@ -281,12 +279,12 @@ internal class StartSatsendringTest {
                 UGYLDIG_SATS
             )
 
-        assertThat(harAlleredeSisteSats(atyMedUglydligSatsUtvidet, SATSTIDSPUNKT)).isEqualTo(false)
+        assertThat(atyMedUglydligSatsUtvidet.erOppdatertMedSatserTilOgMed(SATSTIDSPUNKT)).isEqualTo(false)
 
         val atyMedUgyldigSatsBareOrba =
             lagAndelTilkjentYtelseMedEndreteUtbetalinger(SatsType.ORBA, behandling, ORDINÆR_BARNETRYGD, UGYLDIG_SATS)
 
-        assertThat(harAlleredeSisteSats(atyMedUgyldigSatsBareOrba, SATSTIDSPUNKT)).isEqualTo(false)
+        assertThat(atyMedUgyldigSatsBareOrba.erOppdatertMedSatserTilOgMed(SATSTIDSPUNKT)).isEqualTo(false)
 
         val atyMedUgyldigSatsTilleggOrba =
             lagAndelTilkjentYtelseMedEndreteUtbetalinger(
@@ -296,7 +294,7 @@ internal class StartSatsendringTest {
                 UGYLDIG_SATS
             )
 
-        assertThat(harAlleredeSisteSats(atyMedUgyldigSatsTilleggOrba, SATSTIDSPUNKT)).isEqualTo(false)
+        assertThat(atyMedUgyldigSatsTilleggOrba.erOppdatertMedSatserTilOgMed(SATSTIDSPUNKT)).isEqualTo(false)
     }
 
     @Test
@@ -316,8 +314,7 @@ internal class StartSatsendringTest {
             lagAndelTilkjentYtelseMedEndreteUtbetalinger(SatsType.ORBA, behandling, ORDINÆR_BARNETRYGD)
 
         assertThat(
-            harAlleredeSisteSats(
-                atyMedBGyldigOrba + atyMedGyldigUtvidet + atyMedUgyldigSatsSmåbarnstillegg,
+            (atyMedBGyldigOrba + atyMedGyldigUtvidet + atyMedUgyldigSatsSmåbarnstillegg).erOppdatertMedSatserTilOgMed(
                 SATSTIDSPUNKT
             )
         ).isEqualTo(false)
@@ -339,8 +336,7 @@ internal class StartSatsendringTest {
             )
 
         assertThat(
-            harAlleredeSisteSats(
-                listOf(atySomGårUtPåSatstidspunktGyldig),
+            listOf(atySomGårUtPåSatstidspunktGyldig).erOppdatertMedSatserTilOgMed(
                 SATSTIDSPUNKT
             )
         ).isEqualTo(true)
@@ -358,8 +354,7 @@ internal class StartSatsendringTest {
             )
 
         assertThat(
-            harAlleredeSisteSats(
-                listOf(atySomGårUtPåSatstidspunktUgyldig),
+            listOf(atySomGårUtPåSatstidspunktUgyldig).erOppdatertMedSatserTilOgMed(
                 SATSTIDSPUNKT
             )
         ).isEqualTo(false)
@@ -381,8 +376,7 @@ internal class StartSatsendringTest {
             )
 
         assertThat(
-            harAlleredeSisteSats(
-                listOf(utgåttAndelTilkjentYtelse),
+            listOf(utgåttAndelTilkjentYtelse).erOppdatertMedSatserTilOgMed(
                 SATSTIDSPUNKT
             )
         ).isEqualTo(true)
@@ -404,8 +398,7 @@ internal class StartSatsendringTest {
             )
 
         assertThat(
-            harAlleredeSisteSats(
-                listOf(utgåttAndelTilkjentYtelse),
+            listOf(utgåttAndelTilkjentYtelse).erOppdatertMedSatserTilOgMed(
                 SATSTIDSPUNKT
             )
         ).isEqualTo(true)
@@ -427,8 +420,7 @@ internal class StartSatsendringTest {
             )
 
         assertThat(
-            harAlleredeSisteSats(
-                listOf(utgåttAndelTilkjentYtelse),
+            listOf(utgåttAndelTilkjentYtelse).erOppdatertMedSatserTilOgMed(
                 SATSTIDSPUNKT
             )
         ).isEqualTo(false)
